backend/requirements.txt
# development tools
# ruff
# pre-commit~=3.5
# coverage~=7.3
# pytest~=7.4
ipykernel~=6.16
setuptools>=64.0.0

# project packages
fastapi
uvicorn
openai
python-dotenv
python-multipart
fashion-clip
Pillow
notebook  # required for Jupyter
